+++ src/main/java/chess/view/ChessBoardExpression.java
@@ -1,6 +1,6 @@
 package chess.view;
 
-import chess.domain.Piece;
+import chess.domain.piece.Piece;
 import java.util.List;
 import java.util.stream.Collectors;
