+++ src/test/java/chess/domain/PieceTest.java
@@ -1,11 +1,12 @@
 package chess.domain;
 
+import chess.domain.piece.Piece;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static chess.domain.PieceType.KING;
-import static chess.domain.Team.BLACK;
+import static chess.domain.piece.PieceType.KING;
+import static chess.domain.piece.Team.BLACK;
 
 class PieceTest {
